sse client folder renamed
