--- src/main/java/io/split/android/client/network/eventsource/NotificationProcessor.java
@@ -1,4 +1,0 @@
-package io.split.android.client.network.eventsource;
-
-public class NotificationProcessor {
-}
