+++ src/main/java/io/split/android/client/network/sseclient/EventSourceListener.java
@@ -1,4 +1,4 @@
-package io.split.android.client.network.eventsource;
+package io.split.android.client.network.sseclient;
 
 import java.util.Map;
