+++ src/main/java/io/split/android/client/network/sseclient/NotificationParser.java
@@ -1,4 +1,4 @@
-package io.split.android.client.network.eventsource;
+package io.split.android.client.network.sseclient;
 
 import androidx.annotation.VisibleForTesting;
