+++ src/main/java/io/split/android/client/network/sseclient/NotificationProcessor.java
@@ -1,0 +1,4 @@
+package io.split.android.client.network.sseclient;
+
+public class NotificationProcessor {
+}
